construct ClassTypeDescriptors through factory method
Benchmarks of missinglink found that millions of ClassTypeDescriptor
instances are constructed for a test project that has ~50000 classes in
it's "known class" map.

Since ClassTypeDescriptor does String replacement in it's constructor,
this causes a lot of extra String instances to be created as well.

Forcing the class to be created through a factory - and caching the
instances - is intended to cut down on the memory allocation pressure
to reduce memory footprint when running missinglink (and garbage
collection times).