add warning if no classes found in build directory
This is to help suggest to users that they need to run the compiler
before missinglink.

An alternative would be to force the plugin to run in a forked build so
that we can specify that it runs after the `compile` phase, but this
adds time to the user's build and can cause confusion when some plugins
are run twice (once in the regular build and once in the fork).

Resolves #31