add java.lang.invoke.VarHandle to ClassLoader.BLACKLIST
since all of its methods are annotated with
`@HotSpotIntrinsicCandidate`, which means the methods do not actually
appear in the .class file.

I have not added a unit test here since the project still builds with
`source=1.8, target=1.8` in the compiler configuration (and also builds
in CI on Github with Java 8 as one of the java versions to build with).
The VarHandle class was added in JDK9 and does not appear in 8, so to
add a test of class that references VarHandle I'd have to either resort
to some reflection trickery or drop support in the project for building
with `target=8`, which I don't want to do at this time.

I also considered adding logic to exclude all calls to methods annotated
with `@HotSpotIntrinsicCandidate` from missinglink's analysis/warnings,
but this would adding information about a method's annotations to
missinglink's data model (e..g in DeclaredMethod) which does not exist
today.